@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ToastrModule } from 'ngx-toastr';
@@ -10,23 +10,21 @@ import { LoggerService } from './services/logger.service';
 import { ImportantBookingInfoModule } from './shared/components/important-booking-info/important-booking-info.module';
 import { TableTemplateModule } from './shared/components/table-template/table-template.module';
 import { ConfigService } from './shared/services/config.service';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule,
+    declarations: [
+        AppComponent, HeaderComponent, FooterComponent
+    ],
+    imports: [RouterTestingModule,
         RegistrationModule,
-        HttpClientTestingModule,
         ImportantBookingInfoModule,
         TableTemplateModule,
-        ToastrModule.forRoot()
-      ],
-      declarations: [
-        AppComponent, HeaderComponent, FooterComponent
-      ],
-      providers: [ConfigService, LoggerService]
-    }).compileComponents();
+        ToastrModule.forRoot()],
+    providers: [ConfigService, LoggerService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+}).compileComponents();
   });
 
   it('should create the app', () => {
